codec2: CB2 avi file played stuck [1/2]
PD#OTT-59741

Problem:
The PTS of AVI files does not increase continuously and uniformly. We
cannot simply use the FLAG of UNSTABLE_PTS to incrementally calculate
the PTS.
Because most videos in AVI contain DiscardFlag, this will cause some
frames to need to display multiple FrameDurations.

Solution:
Set the new Flag to codec2 and use the new logic to calculate the PTS of
the AVI video.

Verify:
Android U

Change-Id: Ie8fa9459f4c11758ced87d9ececec21c63d5c9bd
Signed-off-by: liang.hou <liang.hou@amlogic.com>